pyproject.toml
[project]
name = "a11yhood-backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "aiosqlite>=0.22.0",
    "fastapi>=0.124.4",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "requests-oauthlib>=2.0.0",
    "sqlalchemy>=2.0.45",
    "supabase>=2.27.0",
    "uvicorn[standard]>=0.38.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-watch>=4.2.0",
]
